Rename pie chart element
