+++ plugins/cncf-stats/src/components/Charts/StyledPieChart.tsx
